spring cloud gateway의 yml 파일에 routes 설정을 java config로 이관
