+++ apigateway-service/src/main/resources/application.yml
@@ -11,14 +11,15 @@ eureka:
 spring:
   application:
     name: apigateway-service
-  cloud:
-    gateway:
-      routes:
-        - id: first-service
-          uri: http://127.0.0.1:8081/
-          predicates:
-            - Path=/first-service/**
-        - id: second-service
-          uri: http://127.0.0.1:8082/
-          predicates:
-            - Path=/second-service/**
+## -- java code로 이관
+#  cloud:
+#    gateway:
+#      routes:
+#        - id: first-service
+#          uri: http://127.0.0.1:8081/
+#          predicates:
+#            - Path=/first-service/**
+#        - id: second-service
+#          uri: http://127.0.0.1:8082/
+#          predicates:
+#            - Path=/second-service/**
