+++ my-first-service/src/main/java/inflern/study/myfirstservice/MyFirstServiceController.java
@@ -1,0 +1,24 @@
+package inflern.study.myfirstservice;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/first-service")
+@Slf4j
+public class MyFirstServiceController {
+
+    @GetMapping("/welcome")
+    public String welcome() {
+        return "Welcome to the First service.";
+    }
+
+    @GetMapping("/message")
+    public String message(@RequestHeader("first-request") String header) {
+        log.info(header);
+        return "Hello World in First Service";
+    }
+}
