+++ my-second-service/src/main/java/inflern/study/mysecondservice/MySecondServiceController.java
@@ -1,0 +1,24 @@
+package inflern.study.mysecondservice;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/second-service")
+@Slf4j
+public class MySecondServiceController {
+
+    @GetMapping("/welcome")
+    public String welcome() {
+        return "Welcome to the Second service.";
+    }
+
+    @GetMapping("/message")
+    public String message(@RequestHeader("second-request") String header) {
+        log.info(header);
+        return "Hello World in Second Service";
+    }
+}
